keep the .svn dirs on disc during the dev cycle
git-svn-id: http://svn.osgeo.org/osgeo/livedvd/gisvm/trunk@11461 7da46fd8-bd2a-0410-a152-cf9810cf5b10

+++ bin/inchroot_nightly.sh
@@ -162,9 +162,13 @@ export USER_NAME
 
 # save space on ISO by removing the .svn/ dirs
 #   (or control this in bootstrap.sh by uncommenting the 'svn export' line)
-for DIR in `find /usr/local/share/gisvm | grep '\.svn$'` ; do
-   rm -rf "$DIR"
-done
+NEAR_RC=0
+if [ "$NEAR_RC" -eq 1 ] ; then
+   for DIR in `find /usr/local/share/gisvm | grep '\.svn$'` ; do
+      rm -rf "$DIR"
+   done
+fi
+
 # user shouldn't own outside of /home
 chown -R root.root /usr/local/share/gisvm
